+++ src/main/java/com/baedalping/delivery/domain/user/dto/response/UserReadResponseDto.java
@@ -1,4 +1,4 @@
-package com.baedalping.delivery.domain.user.dto;
+package com.baedalping.delivery.domain.user.dto.response;
 
 import com.baedalping.delivery.domain.user.entity.User;
 import java.time.LocalDateTime;
@@ -6,48 +6,60 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
 @NoArgsConstructor
-public class UserCreateResponseDto {
+@Getter
+public class UserReadResponseDto{
   private Long userId;
   private String userName;
   private String email;
   private String userRole;
   private boolean isPublic;
   private LocalDateTime createdAt;
+  private String createBy;
   private LocalDateTime updatedAt;
+  private String updatedBy;
   private LocalDateTime deletedAt;
+  private String deletedBy;
 
   @Builder
-  private UserCreateResponseDto(
+  private UserReadResponseDto(
       Long userId,
       String userName,
       String email,
       String userRole,
       boolean isPublic,
       LocalDateTime createdAt,
+      String createBy,
       LocalDateTime updatedAt,
-      LocalDateTime deletedAt) {
+      String updatedBy,
+      LocalDateTime deletedAt,
+      String deletedBy) {
     this.userId = userId;
     this.userName = userName;
     this.email = email;
     this.userRole = userRole;
     this.isPublic = isPublic;
     this.createdAt = createdAt;
+    this.createBy = createBy;
     this.updatedAt = updatedAt;
+    this.updatedBy = updatedBy;
     this.deletedAt = deletedAt;
+    this.deletedBy = deletedBy;
   }
 
-  public static UserCreateResponseDto ofEntity(User user) {
-    return UserCreateResponseDto.builder()
+  public static UserReadResponseDto ofEntity(User user) {
+    return UserReadResponseDto.builder()
         .userId(user.getUserId())
         .userName(user.getUsername())
         .email(user.getEmail())
         .userRole(user.getRole().getRoleName())
         .isPublic(user.isPublic())
         .createdAt(user.getCreatedAt())
+        .createBy(user.getCreatedBy())
         .updatedAt(user.getUpdatedAt())
+        .updatedBy(user.getUpdatedBy())
         .deletedAt(user.getDeletedAt())
+        .deletedBy(user.getDeletedBy())
         .build();
   }
 }
